fixing regression introduced by 05a3b2 - now that mtime was properly saved and it was now used to avoid full saves() (#429) it is skipping over parse_persistent_info and also job status checking logic.
introduced check_job_status() to CachedClusterObject.  Will update the object, the related Job, and return a dictionary of changes.  This allows either save() or query.update() to be used to update the dataabase depending on a full save or update query

we will likely want to move this logic, including last_job field, into CachedClusterObject.  It should be fairly easy since the implementation in VirtualMachine is generic.