Migrations: updating migrations to be easier to automatic for upgrades from 0.6 or earlier:
   since the ganeti app was renamed south doesn't know what migration a user was at.  That information is stored with the ganeti app.
That would mean the admin would have to fake apply 6 migration files.  That will be too complicated for many people when it can be avoided.

The solution:
   * create app deprecated.ganeti that contains just the migrations from 0.4 through 0.6
   * update 0.7 migrations to use ganeti_web for all tables
   * add migrations for 0.7 that will convert values from 0.6 tables if they exist, and then delete the 0.6 tables

note that the deprecated.ganeti namespace was chosen because it means a reference to "ganeti" will still not work, but django will still
treat it as the old "ganeti".   django only uses the last package in the app path as the name.