Tickets #7059 & #7065:    * Display all running/queued jobs on an object instead of just last job    * display jobs started by other users on detail screens
This is a rework of the javascript and backend view that reports job info.  The changes that it introduces are:
   * ajax call queries all running,waiting, and non-cleared errors.
   * rendering code has been updated to reuse and update html elements when already present
   * job polling always runs.  When no jobs have been returned it uses a slower polling interval
   * init and interval poll have been separated into separate methods.  This is to better support moving this code into a closure in another commit.
   * running, waiting, error'd jobs are now marked up appropriately.
   * VirtualMachine, Node, and Cluster pages have all been updated.

NOTE: that this code will still use existing callback methods which currently refresh the page after each job finishes.  There is another ticket to deal with processing jobs without refreshing the page (#3255)