Additions to slider functionality for feature 6597
Sliders now function as intended after an error on VM create,
but are still not present on the modify or template pages.
Also, the max and min values for the sliders are still
hardcoded, rather than being set by a query for cluster quotas.
Finally, all sliders are currently using powers of 2
(MiB, GiB, etc) rather than powers of 10 (MB, GB, etc).
These things will likely be changed in a future commit.

refs #6597