views/user: #9987: Rewrite UserEditForm for Django 1.4.
Remove the password change functionality; it was duplicated between two views
and doesn't work in Django 1.4. Also adjust the fields slightly to account for
a quirk in UserChangeForm's implementation.

Also nuke tests for the removed functionality. Tests for the view which
changes passwords are still present and working. Actually, all tests in
muddle_users still pass with this change.

Fixes #9987.