fields: Rewrite, simplify, and test PDTField.
Man, some of this stuff is icky.

Should be tested on My and Pg, but there's no reason this wouldn't work: The
db_type() stuff did *exactly* what was already done in each individual DB
backend already. There's a good chance this unbreaks Oracle. (As if we cared!)

Originally from a stashed patch from the 0.8 cycle. I figured I would port and
share it with everybody else.