Ticket #9093 - Fixes sorting of cluster version.
Tablesorter is pretty dumb and interprets our ganeti
version as a float. When it does this it truncates
the version number when it finds a second period.
Adding a 'v' before the version number makes table-
sorter interpret the number as a text string instead.

This is a simpler fix than telling the field to use a
custom parser that we would have to write and maintain.

fixes #9093