backend/queries, views: Factor out another query.
This query was buggy in one spot in views/vms, causing some VMs to not show up
consistently on all pages.

Additionally, the query in views/general was inconsistently picking either a
query of VMs or a query of VM PKs, which could have been a type problem.

By factoring out queries this way, we not only reduce code duplication, but
also clear away the possibility of bugs and make code easier to test.

Refs #12171.