Added a lot of placeholder docs with TODOs.
This includes quite a bit of rearranging on the index page.

Added Usage section back, so there now exists both Usage and Feature
documentation.

conf.py now contains the default setting for todo_include_todos. This
can be set to True to generate documentation that displays todos. It
will not be picked up by sphinx-build though, so the build directory
will have to be rm-ed first.

Refs #12279