Fix Ctrl-Alt-Del confirmation in noVNC
This removes the Ctrl-Alt-Del button from the power class (where the
previous, broken confirmation dialog box was from) and puts a different
confirmation box in the noVNC.js file.

The power class javascript did nothing for the Ctrl-Alt-Del button,
since the javascript in novnc.html is what sends the Ctrl-Alt-Del.
Therefore, the dialog must go in novnc.html or the Ctrl-Alt-Del send
must be moved out of novnc.html.

References #16377