Fix AUTH_PROFILE_MODULE path
There are no gwm docs for what the hell this thing does so this is my best
stab at what it should be set to.

AUTH_PROFILE_MODULE is a setting which tells Django where your user model is.
Ours seems to be in `ganeti_webmgr/authentication/models.py` in a class called
`Profile`.

Interestingly enough the cookbook doesn't set this.