Bump requests version
This should address #28 on GitHub. The requests API has not changed and the new
version of requests has a newer urllib3 library.
I pinned the version to >=2.7.0 because I don't think that breaking changes
will be made to this package. Frankly I don't know why this package isn't in
the standard library.