docker-compose development environment
This adjusts the development environment to use docker-compose which allows us
to spin up a "fake" three node ganeti cluster with four instances "running".
The clusters are setup using a fake hypervisor so it's not a complete
replacement for developing against a real cluster, but it's as close as we can
get.

This also provides updates for the following:

- Cleaned up the Dockerfile and moved parts of the work into a script
- Switched to using a volume for /opt/ganeti_webmgr which allows for easier editing
- Switched to using osuosl/django-centos image which already has stuff
  pre-included